requirements.txt
django==4.1.7
django-countries==7.5.1
numpy==1.23.5
Pillow==9.4.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
django-notifications-hq==1.7.0
django-ckeditor==6.5.1
django-allauth==0.53.1
djangorestframework==3.14.0
schedule==1.2.0
selenium==4.9.1
scrapy==2.9.0
beautifulsoup4==4.12.2
requests==2.28.2
pandas==2.0.1
social-auth-app-django==5.1.0
stripe==5.4.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
scikit-learn==1.2.2
